requirements-dev.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
black==24.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.0.0
fonttools==4.58.4
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.8
limits==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
psutil==7.0.0
py-cpuinfo==9.0.0
pycodestyle==2.11.1
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.0.2
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.155
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
freezegun==1.5.1
